refactor(terminal): single-source the copy/paste shortcut matrix
Address CodeRabbit review on #1204:
- Move decideTerminalKeyAction into src/shared/terminalShortcuts.ts so the
  renderer console and the injected stopgap terminal share one source. The
  injected script now embeds the shared function via .toString() instead of
  keeping a hand-maintained copy, so the two can't drift.
- Swallow the async clipboard writeText() rejection in the injected copy
  path (not just sync throws).

Amp-Thread-ID: https://ampcode.com/threads/T-019f0347-5c56-713e-9895-f12f03c04ac0
